07-rest-api-docs/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["CristiPV <cristi.valentin29@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.82.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
python-dotenv = "^0.21.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
